refactor(e2e): decompose llm_judge.py from 1,488 to 664 lines
Extract three focused modules from the monolithic llm_judge.py:
- llm_judge_models.py: JudgeResult and BuildPipelineResult data models
- build_pipeline.py: Build/lint/test pipeline execution functions
- pipeline_scripts.py: Script creation and log saving functions

All existing imports preserved via re-exports in llm_judge.py.

Closes #1359
